api-gateway/requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
flask-cors==2.1.3
PyJWT==2.1.0
gunicorn==20.1.0 